Only set the Parent-Request-Id when not nil
Why you made the change:

I made this change because Rack linting raises an exception when it gets
a header that is nil. This is because it is expects headers to be
strings or to not exist if they have no value as that is generally what
is done in the HTTP spec.

How the change addresses the need:

This change makes it so that the header isn't included when it is nil.
Making it abide by the standard and preventing Rack linting from
raising an exception.